api/requirements.txt
fastapi
uvicorn
langchain-community
langchain-core
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
python-dotenv
reportlab
